Розделил код юзера
