+++ controllers/user/index.js
@@ -1,5 +1,9 @@
 const logout = require('./logout');
+const signup = require('./signup');
+const login = require('./login');
 
 module.exports = {
   logout,
+  signup,
+  login,
 };
