+++ controllers/user/signup.js
@@ -1,0 +1,33 @@
+const { BadRequest, Conflict } = require('http-errors');
+const bcrypt = require('bcryptjs');
+
+const { User } = require('../../models');
+const { joiRegisterSchema } = require('../../models/user');
+
+const signup = async (req, res, next) => {
+  try {
+    const { error } = joiRegisterSchema.validate(req.body);
+    if (error) {
+      throw new BadRequest(error.message);
+    }
+
+    const { name, email, password } = req.body;
+    const user = await User.findOne({ email });
+    if (user) {
+      throw new Conflict('Email in use');
+    }
+    const salt = await bcrypt.genSalt(10);
+    const hashPassword = await bcrypt.hash(password, salt);
+    const newUser = await User.create({ name, email, password: hashPassword });
+    res.status(201).json({
+      user: {
+        name: newUser.name,
+        email: newUser.email,
+      },
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = signup;
